cambios que no afectan la funcionalidad, solo una validacion con js para el registro

+++ luzazul/crudProductos.php
@@ -16,7 +16,7 @@ include_once('partials/header.php');
       <section class="col- 12 col-md-12">
         <h1 class="col-12 text-danger text-center">Administrar productos</h1>
         <button class="btn btn-primary col-12 col-md-2 offset-md-5"><a class="text-white" href="agregarProducto.php">Agregar Producto</a></button>
-        <table class="table col-12 col-md-8 offset-md-2 mt-5">
+        <table class="table col-12 col-md-8 offset-md-2 mt-5 table-hover">
           <thead class="col-12">
             <tr>
               <th class="text-danger" scope="col">#</th>
@@ -39,7 +39,7 @@ include_once('partials/header.php');
                 <td class="text-white"><a href="editarProducto.php?id=<?= $value['id']; ?>">
                     <ion-icon name="build-outline"></ion-icon>
                   </a></td>
-                <td class="text-white"><a href="eliminarProducto.php?id=<?= $value['id']; ?>">
+                <td class="text-white"><a onclick="return confirm('¿Estas seguro que queres eliminar el proyecto?')" href="eliminarProducto.php?id=<?= $value['id']; ?>">
                     <ion-icon name="trash-outline"></ion-icon>
                   </a></td>
 
@@ -63,7 +63,7 @@ include_once('partials/header.php');
         <?php
         include_once('partials/nav.php');
         ?>
-        <section class="col- 12 col-md-12">
+        <section class="col- 12 col-md-12 bg-danger">
           <h1 class="col-12 text-center">No se puede acceder a este sitio.</h1>
         </section>
         <?php
@@ -77,7 +77,7 @@ include_once('partials/header.php');
         <?php
         include_once('partials/nav.php');
         ?>
-        <section class="col- 12 col-md-12">
+        <section class="col- 12 col-md-12 bg-danger">
           <h1 class="col-12 text-center">No se puede acceder a este sitio.</h1>
         </section>
         <?php
